Extract utility methods from McpServerProcessor
- to reduce the number of lines from 2131 to 1607 (~25%)
- also fix a few minor inefficiencies
