+++ core/deployment/src/main/java/io/quarkiverse/mcp/server/deployment/McpServerProcessor.java
@@ -17,8 +17,6 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.Comparator;
 import java.util.HashMap;
-import java.util.HashSet;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
@@ -36,17 +34,14 @@ import jakarta.enterprise.invoke.Invoker;
 import jakarta.inject.Singleton;
 
 import org.jboss.jandex.AnnotationInstance;
-import org.jboss.jandex.AnnotationTarget;
 import org.jboss.jandex.AnnotationValue;
 import org.jboss.jandex.ClassInfo;
 import org.jboss.jandex.ClassInfo.NestingType;
 import org.jboss.jandex.ClassType;
 import org.jboss.jandex.DotName;
-import org.jboss.jandex.EquivalenceKey.TypeEquivalenceKey;
 import org.jboss.jandex.IndexView;
 import org.jboss.jandex.MethodInfo;
 import org.jboss.jandex.MethodParameterInfo;
-import org.jboss.jandex.PrimitiveType;
 import org.jboss.jandex.Type.Kind;
 import org.jboss.jandex.gizmo2.Jandex2Gizmo;
 import org.jboss.logging.Logger;
@@ -58,7 +53,6 @@ import io.quarkiverse.mcp.server.CompleteArg;
 import io.quarkiverse.mcp.server.Content;
 import io.quarkiverse.mcp.server.DefaultValueConverter;
 import io.quarkiverse.mcp.server.EmbeddedResource;
-import io.quarkiverse.mcp.server.ExecutionModel;
 import io.quarkiverse.mcp.server.GlobalInputSchemaGenerator;
 import io.quarkiverse.mcp.server.GlobalOutputSchemaGenerator;
 import io.quarkiverse.mcp.server.Icon;
@@ -113,7 +107,6 @@ import io.quarkiverse.mcp.server.runtime.ResourceContentsEncoderResultMapper;
 import io.quarkiverse.mcp.server.runtime.ResourceManagerImpl;
 import io.quarkiverse.mcp.server.runtime.ResourceTemplateCompletionManagerImpl;
 import io.quarkiverse.mcp.server.runtime.ResourceTemplateManagerImpl;
-import io.quarkiverse.mcp.server.runtime.ResourceTemplateManagerImpl.VariableMatcher;
 import io.quarkiverse.mcp.server.runtime.ResultMappers;
 import io.quarkiverse.mcp.server.runtime.SamplingRequestImpl;
 import io.quarkiverse.mcp.server.runtime.ServerRequests;
@@ -149,7 +142,6 @@ import io.quarkus.arc.processor.BeanInfo;
 import io.quarkus.arc.processor.BuiltinScope;
 import io.quarkus.arc.processor.InjectionPointInfo;
 import io.quarkus.arc.processor.InvokerBuilder;
-import io.quarkus.arc.processor.KotlinUtils;
 import io.quarkus.arc.processor.RuntimeTypeCreator;
 import io.quarkus.deployment.GeneratedClassGizmo2Adaptor;
 import io.quarkus.deployment.annotations.BuildProducer;
@@ -187,8 +179,6 @@ import io.vertx.core.json.Json;
 class McpServerProcessor {
 
     private static final Logger LOG = Logger.getLogger(McpServerProcessor.class);
-
-    private static final String DEFAULT_VALUE = "defaultValue";
 
     @BuildStep
     FeatureAnnotationsBuildItem featureAnnotations(McpServersBuildTimeConfig config) {
@@ -305,7 +295,8 @@ class McpServerProcessor {
             BuildProducer<FeatureMethodBuildItem> features,
             BuildProducer<ValidationErrorBuildItem> errors) {
 
-        List<Throwable> wrongUsages = findWrongAnnotationUsage(beanArchiveIndex.getIndex(), featureAnnotations, errors);
+        List<Throwable> wrongUsages = FeatureMethods.findWrongAnnotationUsage(beanArchiveIndex.getIndex(),
+                featureAnnotations, errors);
         if (!wrongUsages.isEmpty()) {
             errors.produce(new ValidationErrorBuildItem(wrongUsages));
         }
@@ -318,8 +309,8 @@ class McpServerProcessor {
                 AnnotationInstance featureAnnotation = featureAnnotations.getFeatureAnnotation(method);
                 if (featureAnnotation != null) {
                     Feature feature = featureAnnotations.getFeature(featureAnnotation);
-                    validateFeatureMethod(method, feature, featureAnnotation, defaultValueConverters,
-                            beanArchiveIndex.getIndex());
+                    FeatureMethods.validateFeatureMethod(method, feature, featureAnnotation,
+                            defaultValueConverters, beanArchiveIndex.getIndex());
                     String name;
                     if (feature == PROMPT_COMPLETE
                             || feature == RESOURCE_TEMPLATE_COMPLETE) {
@@ -447,7 +438,7 @@ class McpServerProcessor {
                     }
 
                     // @McpServer bindings
-                    Set<String> servers = initServerBindings(config, beanArchiveIndex.getIndex(), method);
+                    Set<String> servers = FeatureMethods.initServerBindings(config, beanArchiveIndex.getIndex(), method);
 
                     // Process metadata entries
                     AnnotationInstance metaField = method.declaredAnnotation(DotNames.META_FIELD);
@@ -533,7 +524,8 @@ class McpServerProcessor {
                             servers, structuredContent,
                             outputSchemaFrom, outputSchemaGenerator, inputSchemaGenerator, resourceAnnotations,
                             metadata,
-                            inputGuardrails, outputGuardrails, executionModel(method, transformedAnnotations),
+                            inputGuardrails, outputGuardrails,
+                            FeatureMethods.executionModel(method, transformedAnnotations),
                             iconsProvider);
                     features.produce(fm);
                     found.compute(feature, (f, list) -> {
@@ -625,72 +617,6 @@ class McpServerProcessor {
         }
     }
 
-    private Set<String> initServerBindings(McpServersBuildTimeConfig config, IndexView index, MethodInfo method) {
-        Set<String> ret = new HashSet<String>();
-        Optional<Boolean> multiServerBindings = config.supportMultiServerBindings();
-        if (multiServerBindings.orElse(true)) {
-            // The set of bindings includes all values declared on the feature method and
-            // all values defined on the declaring class of the feature
-            List<AnnotationInstance> methodAnnotations = new ArrayList<>(
-                    method.declaredAnnotationsWithRepeatable(DotNames.MCP_SERVER, index));
-            methodAnnotations.addAll(method.declaredAnnotationsWithRepeatable(DotNames.MCPJAVA_MCP_SERVER, index));
-            List<AnnotationInstance> classAnnotations = new ArrayList<>(
-                    method.declaringClass().declaredAnnotationsWithRepeatable(DotNames.MCP_SERVER, index));
-            classAnnotations.addAll(
-                    method.declaringClass().declaredAnnotationsWithRepeatable(DotNames.MCPJAVA_MCP_SERVER, index));
-            for (AnnotationInstance a : methodAnnotations) {
-                ret.add(a.value().asString());
-            }
-            for (AnnotationInstance a : classAnnotations) {
-                ret.add(a.value().asString());
-            }
-            if (ret.size() > 1 && methodAnnotations.size() == 1 && classAnnotations.size() == 1
-                    && multiServerBindings.isEmpty()) {
-                // In versions 1.10 and lower, the method-level annotation overrode the
-                // class-level one;
-                // now both are merged - fail the build if the user has not explicitly opted in
-                throw new IllegalStateException(
-                        String.format(
-                                "Feature method %s#%s() is bound to servers %s because @McpServer bindings declared on"
-                                        + " the method and the declaring class are now merged."
-                                        + " In previous versions, the method-level binding would override the class-level one."
-                                        + " Set 'quarkus.mcp.server.support-multi-server-bindings=true' to use the new merge behavior,"
-                                        + " or 'quarkus.mcp.server.support-multi-server-bindings=false' to revert to the old behavior.",
-                                method.declaringClass().name().withoutPackagePrefix(), method.name(), ret));
-            }
-        } else {
-            // Compatibility mode - only a single binding is allowed
-            List<AnnotationInstance> methodAnnotations = new ArrayList<>(
-                    method.declaredAnnotationsWithRepeatable(DotNames.MCP_SERVER, index));
-            methodAnnotations.addAll(method.declaredAnnotationsWithRepeatable(DotNames.MCPJAVA_MCP_SERVER, index));
-            if (methodAnnotations.size() > 1) {
-                throw new IllegalStateException(
-                        "Only single @McpServer binding is allowed in compatibility mode: " + method.declaringClass()
-                                + "#"
-                                + method.name()
-                                + "()");
-            } else if (methodAnnotations.size() == 1) {
-                ret.add(methodAnnotations.get(0).value().asString());
-            } else {
-                // Try the declaring class
-                List<AnnotationInstance> classAnnotations = new ArrayList<>(
-                        method.declaringClass().declaredAnnotationsWithRepeatable(DotNames.MCP_SERVER, index));
-                classAnnotations.addAll(
-                        method.declaringClass().declaredAnnotationsWithRepeatable(DotNames.MCPJAVA_MCP_SERVER, index));
-                if (classAnnotations.size() > 1) {
-                    throw new IllegalStateException(
-                            "Only single @McpServer binding is allowed in compatibility mode: "
-                                    + method.declaringClass() + "#"
-                                    + method.name()
-                                    + "()");
-                } else if (classAnnotations.size() == 1) {
-                    ret.add(classAnnotations.get(0).value().asString());
-                }
-            }
-        }
-        return ret.isEmpty() ? Set.of(McpServer.DEFAULT) : Set.copyOf(ret);
-    }
-
     @BuildStep
     void validateGuardrailsAndIconProviders(BeanArchiveIndexBuildItem beanArchiveIndex,
             List<FeatureMethodBuildItem> featureMethods,
@@ -739,24 +665,22 @@ class McpServerProcessor {
                         beans);
                 errors.produce(new ValidationErrorBuildItem(new IllegalStateException(message)));
             } else if (beans.isEmpty()) {
-                if (clazz != null) {
-                    if (clazz.nestingType() == NestingType.INNER && !Modifier.isStatic(clazz.flags())) {
+                if (clazz.nestingType() == NestingType.INNER && !Modifier.isStatic(clazz.flags())) {
+                    errors.produce(new ValidationErrorBuildItem(new IllegalStateException(
+                            componentType + " implementations must not be inner classes: "
+                                    + clazz)));
+                } else {
+                    MethodInfo noArgsConstructor = clazz.method("<init>");
+                    if (noArgsConstructor == null || !Modifier.isPublic(noArgsConstructor.flags())) {
                         errors.produce(new ValidationErrorBuildItem(new IllegalStateException(
-                                componentType + " implementations must not be inner classes: "
+                                componentType
+                                        + " implementations must be CDI beans, or declare a public no-args constructor: "
                                         + clazz)));
                     } else {
-                        MethodInfo noArgsConstructor = clazz.method("<init>");
-                        if (noArgsConstructor == null || !Modifier.isPublic(noArgsConstructor.flags())) {
-                            errors.produce(new ValidationErrorBuildItem(new IllegalStateException(
-                                    componentType
-                                            + " implementations must be CDI beans, or declare a public no-args constructor: "
-                                            + clazz)));
-                        } else {
-                            // Also register the component for reflection
-                            reflectiveClasses
-                                    .produce(ReflectiveClassBuildItem.builder(clazzName.toString())
-                                            .constructors().build());
-                        }
+                        // Also register the component for reflection
+                        reflectiveClasses
+                                .produce(ReflectiveClassBuildItem.builder(clazzName.toString())
+                                        .constructors().build());
                     }
                 }
             }
@@ -859,37 +783,10 @@ class McpServerProcessor {
         return null;
     }
 
-    private List<Throwable> findWrongAnnotationUsage(IndexView index,
-            FeatureAnnotationsBuildItem featureAnnotations,
-            BuildProducer<ValidationErrorBuildItem> validationErrors) {
-        List<Throwable> wrongUsages = new ArrayList<>();
-        for (Map.Entry<DotName, Feature> e : featureAnnotations.annotationToFeature().entrySet()) {
-            findWrongMethods(index, e.getKey(), e.getValue(), wrongUsages);
-        }
-        return wrongUsages;
-    }
-
-    private void findWrongMethods(IndexView index, DotName annotationName, Feature feature,
-            List<Throwable> wrongUsages) {
-        for (AnnotationInstance annotation : index.getAnnotations(annotationName)) {
-            if (annotation.target().kind() == AnnotationTarget.Kind.METHOD) {
-                if (Modifier.isStatic(annotation.target().asMethod().flags())) {
-                    wrongUsages
-                            .add(new IllegalStateException(feature + " method must not be static: "
-                                    + methodDesc(annotation.target().asMethod())));
-                } else if (annotation.target().asMethod().declaringClass().isInterface()) {
-                    wrongUsages
-                            .add(new IllegalStateException(feature + " method must not be declared on an interface: "
-                                    + methodDesc(annotation.target().asMethod())));
-                }
-            }
-        }
-    }
-
     private String getDuplicateValidationName(FeatureMethodBuildItem featureMethod) {
         if (featureMethod.getFeature() == PROMPT_COMPLETE || featureMethod.getFeature() == RESOURCE_TEMPLATE_COMPLETE) {
             MethodParameterInfo argument = featureMethod.getMethod().parameters().stream()
-                    .filter(p -> providerFrom(p.type()) == Provider.PARAMS).findFirst().orElseThrow();
+                    .filter(p -> FeatureArguments.providerFrom(p.type()) == Provider.PARAMS).findFirst().orElseThrow();
             String argumentName = argument.name();
             AnnotationInstance completeArg = argument.declaredAnnotation(DotNames.COMPLETE_ARG);
             if (completeArg != null) {
@@ -1223,7 +1120,7 @@ class McpServerProcessor {
         boolean generateHolder = false;
         List<MethodParameterInfo> serializedArguments = new ArrayList<>();
         for (MethodParameterInfo param : tool.getMethod().parameters()) {
-            if (providerFrom(param.type()) == Provider.PARAMS) {
+            if (FeatureArguments.providerFrom(param.type()) == Provider.PARAMS) {
                 serializedArguments.add(param);
                 List<AnnotationInstance> annotations = param.declaredAnnotations();
                 if (!annotations.isEmpty()
@@ -1271,11 +1168,11 @@ class McpServerProcessor {
             if (DotNames.LIST.equals(returnType.name()) && returnType.kind() == Kind.PARAMETERIZED_TYPE) {
                 returnType = returnType.asParameterizedType().arguments().get(0);
             }
-            if (isReturnTypeReflectionNeeded(returnType.name())) {
+            if (FeatureMethods.isReturnTypeReflectionNeeded(returnType.name())) {
                 reflectiveHierarchies.produce(ReflectiveHierarchyBuildItem.builder(m.getMethod().returnType()).build());
             }
             for (org.jboss.jandex.Type paramType : m.getMethod().parameterTypes()) {
-                if (isParamTypeReflectionNeeded(paramType)) {
+                if (FeatureMethods.isParamTypeReflectionNeeded(paramType)) {
                     reflectiveHierarchies.produce(ReflectiveHierarchyBuildItem.builder(paramType).build());
                 }
             }
@@ -1284,7 +1181,7 @@ class McpServerProcessor {
 
         // Register all default value converters
         reflectiveClasses.produce(ReflectiveClassBuildItem.builder(defaultValueConverters.stream()
-                .map(DefaultValueConverterBuildItem::getClassName).toList().toArray(String[]::new))
+                .map(DefaultValueConverterBuildItem::getClassName).toArray(String[]::new))
                 .constructors()
                 .build());
 
@@ -1377,160 +1274,6 @@ class McpServerProcessor {
         services.produce(ServiceProviderBuildItem.allProvidersFromClassPath(McpServerSPI.class.getName()));
     }
 
-    private static final Set<DotName> ARG_ANNOTATIONS = Set.of(DotNames.TOOL_ARG, DotNames.PROMPT_ARG,
-            DotNames.COMPLETE_ARG,
-            DotNames.RESOURCE_TEMPLATE_ARG,
-            DotNames.MCPJAVA_TOOL_ARG, DotNames.MCPJAVA_PROMPT_ARG, DotNames.MCPJAVA_COMPLETE_ARG,
-            DotNames.MCPJAVA_RESOURCE_TEMPLATE_ARG);
-
-    private void validateFeatureMethod(MethodInfo method, Feature feature, AnnotationInstance featureAnnotation,
-            List<DefaultValueConverterBuildItem> defaultValueConverters, IndexView index) {
-        if (Modifier.isStatic(method.flags())) {
-            throw new IllegalStateException(feature + " method must not be static: " + methodDesc(method));
-        }
-        if (Modifier.isPrivate(method.flags())) {
-            throw new IllegalStateException(feature + " method must not be private: " + methodDesc(method));
-        }
-        if (method.returnType().kind() == Kind.VOID && feature != NOTIFICATION) {
-            throw new IllegalStateException(feature + " method may not return void: " + methodDesc(method));
-        }
-        for (MethodParameterInfo param : method.parameters()) {
-            Set<DotName> validArgAnnotations = switch (feature) {
-                case TOOL -> Set.of(DotNames.TOOL_ARG, DotNames.MCPJAVA_TOOL_ARG);
-                case PROMPT -> Set.of(DotNames.PROMPT_ARG, DotNames.MCPJAVA_PROMPT_ARG);
-                case RESOURCE_TEMPLATE ->
-                    Set.of(DotNames.RESOURCE_TEMPLATE_ARG, DotNames.MCPJAVA_RESOURCE_TEMPLATE_ARG);
-                case PROMPT_COMPLETE, RESOURCE_TEMPLATE_COMPLETE ->
-                    Set.of(DotNames.COMPLETE_ARG, DotNames.MCPJAVA_COMPLETE_ARG);
-                default -> Set.of();
-            };
-            Set<DotName> invalidAnnotations = new HashSet<>(ARG_ANNOTATIONS);
-            invalidAnnotations.removeAll(validArgAnnotations);
-            for (DotName invalidAnnotation : invalidAnnotations) {
-                if (param.hasDeclaredAnnotation(invalidAnnotation)) {
-                    throw new IllegalStateException("Parameter of a %s method may not be annotated with @%s: %s"
-                            .formatted(feature, invalidAnnotation.withoutPackagePrefix(), methodDesc(method)));
-                }
-            }
-        }
-        switch (feature) {
-            case PROMPT -> validatePromptMethod(method);
-            case PROMPT_COMPLETE -> validatePromptCompleteMethod(method);
-            case TOOL -> validateToolMethod(method, defaultValueConverters, index);
-            case RESOURCE -> validateResourceMethod(method);
-            case RESOURCE_TEMPLATE -> validateResourceTemplateMethod(method, featureAnnotation);
-            case RESOURCE_TEMPLATE_COMPLETE -> validateResourceTemplateCompleteMethod(method);
-            case NOTIFICATION -> validateNotificationMethod(method);
-            default -> throw new IllegalArgumentException("Unsupported feature: " + feature);
-        }
-    }
-
-    private static final Set<org.jboss.jandex.Type> PROMPT_TYPES = Set.of(ClassType.create(DotNames.PROMPT_RESPONSE),
-            ClassType.create(DotNames.PROMPT_MESSAGE), ClassType.create(DotNames.STRING));
-
-    private void validatePromptMethod(MethodInfo method) {
-        // No need to validate return type
-
-        List<MethodParameterInfo> parameters = parameters(method, PROMPT);
-        for (MethodParameterInfo param : parameters) {
-            if (!param.type().name().equals(DotNames.STRING)) {
-                throw new IllegalStateException(
-                        "Prompt method must only consume String parameters: " + methodDesc(method));
-            }
-        }
-    }
-
-    private static final Set<org.jboss.jandex.Type> COMPLETE_TYPES = Set.of(
-            ClassType.create(DotNames.COMPLETE_RESPONSE),
-            ClassType.create(DotNames.STRING),
-            ClassType.create(DotNames.MCPJAVA_COMPLETION_RESULT));
-
-    private void validatePromptCompleteMethod(MethodInfo method) {
-        org.jboss.jandex.Type type = method.returnType();
-        if (DotNames.UNI.equals(type.name()) && type.kind() == Kind.PARAMETERIZED_TYPE) {
-            type = type.asParameterizedType().arguments().get(0);
-        }
-        if (DotNames.LIST.equals(type.name()) && type.kind() == Kind.PARAMETERIZED_TYPE) {
-            type = type.asParameterizedType().arguments().get(0);
-        }
-        if (!COMPLETE_TYPES.contains(type)) {
-            throw new IllegalStateException("Unsupported Prompt complete method return type: " + methodDesc(method));
-        }
-
-        List<MethodParameterInfo> parameters = parameters(method, PROMPT_COMPLETE);
-        if (parameters.size() != 1 || !parameters.get(0).type().name().equals(DotNames.STRING)) {
-            throw new IllegalStateException(
-                    "Prompt complete must consume exactly one String argument: " + methodDesc(method));
-        }
-    }
-
-    private void validateResourceTemplateCompleteMethod(MethodInfo method) {
-        org.jboss.jandex.Type type = method.returnType();
-        if (DotNames.UNI.equals(type.name()) && type.kind() == Kind.PARAMETERIZED_TYPE) {
-            type = type.asParameterizedType().arguments().get(0);
-        }
-        if (DotNames.LIST.equals(type.name()) && type.kind() == Kind.PARAMETERIZED_TYPE) {
-            type = type.asParameterizedType().arguments().get(0);
-        }
-        if (!COMPLETE_TYPES.contains(type)) {
-            throw new IllegalStateException(
-                    "Unsupported Resource template complete method return type: " + methodDesc(method));
-        }
-
-        List<MethodParameterInfo> parameters = parameters(method, RESOURCE_TEMPLATE_COMPLETE);
-        if (parameters.size() != 1 || !parameters.get(0).type().name().equals(DotNames.STRING)) {
-            throw new IllegalStateException(
-                    "Resource template complete must consume exactly one String argument: " + methodDesc(method));
-        }
-    }
-
-    private static final Set<org.jboss.jandex.Type> TOOL_TYPES = Set.of(ClassType.create(DotNames.TOOL_RESPONSE),
-            ClassType.create(DotNames.CONTENT), ClassType.create(DotNames.TEXT_CONTENT),
-            ClassType.create(DotNames.IMAGE_CONTENT), ClassType.create(DotNames.EMBEDDED_RESOURCE),
-            ClassType.create(DotNames.AUDIO_CONTENT), ClassType.create(DotNames.RESOURCE_LINK),
-            ClassType.create(DotNames.STRING));
-
-    private void validateToolMethod(MethodInfo method, List<DefaultValueConverterBuildItem> defaultValueConverters,
-            IndexView index) {
-        parameters(method, TOOL);
-        for (MethodParameterInfo p : method.parameters()) {
-            AnnotationInstance toolArg = p.annotation(DotNames.TOOL_ARG);
-            if (toolArg == null) {
-                toolArg = p.annotation(DotNames.MCPJAVA_TOOL_ARG);
-            }
-            if (toolArg != null) {
-                AnnotationValue defaultValueValue = toolArg.value(DEFAULT_VALUE);
-                if (defaultValueValue != null) {
-                    // Strings and enums are handled specifically
-                    if (p.type().name().equals(DotNames.STRING)) {
-                        continue;
-                    }
-                    if (p.type().kind() == Kind.CLASS) {
-                        ClassInfo pclazz = index.getClassByName(p.type().name());
-                        if (pclazz != null && pclazz.isEnum()) {
-                            continue;
-                        }
-                    }
-                    // Make sure primitives are boxed
-                    org.jboss.jandex.Type argType;
-                    if (p.type().kind() == Kind.PRIMITIVE) {
-                        argType = PrimitiveType.box(p.type().asPrimitiveType());
-                    } else {
-                        argType = p.type();
-                    }
-                    TypeEquivalenceKey argEquivalenceKey = TypeEquivalenceKey.of(argType);
-                    if (defaultValueConverters.stream()
-                            .noneMatch(c -> TypeEquivalenceKey.of(c.getArgumentType()).equals(argEquivalenceKey))) {
-                        throw new IllegalStateException(
-                                "No matching default value converter found for argument type [" + p.type()
-                                        + "] declared on: "
-                                        + p);
-                    }
-                }
-            }
-        }
-    }
-
     private boolean useEncoder(org.jboss.jandex.Type type, Set<org.jboss.jandex.Type> types) {
         if (DotNames.UNI.equals(type.name()) && type.kind() == Kind.PARAMETERIZED_TYPE) {
             type = type.asParameterizedType().arguments().get(0);
@@ -1539,84 +1282,6 @@ class McpServerProcessor {
             type = type.asParameterizedType().arguments().get(0);
         }
         return !types.contains(type);
-    }
-
-    static final Set<org.jboss.jandex.Type> RESOURCE_TYPES = Set.of(ClassType.create(DotNames.RESOURCE_RESPONSE),
-            ClassType.create(DotNames.RESOURCE_CONTENTS), ClassType.create(DotNames.TEXT_RESOURCE_CONTENTS),
-            ClassType.create(DotNames.BLOB_RESOURCE_CONTENTS));
-
-    private void validateResourceMethod(MethodInfo method) {
-        // No need to validate return type
-
-        List<MethodParameterInfo> parameters = parameters(method, RESOURCE);
-        if (!parameters.isEmpty()) {
-            throw new IllegalStateException(
-                    "Resource method may only accept built-in parameter types" + methodDesc(method));
-        }
-    }
-
-    private void validateResourceTemplateMethod(MethodInfo method, AnnotationInstance featureAnnotation) {
-        // No need to validate return type
-
-        AnnotationValue uriTemplateValue = featureAnnotation.value("uriTemplate");
-        if (uriTemplateValue == null) {
-            throw new IllegalStateException("URI template not found");
-        }
-        VariableMatcher variableMatcher = ResourceTemplateManagerImpl
-                .createMatcherFromUriTemplate(uriTemplateValue.asString());
-
-        List<MethodParameterInfo> parameters = parameters(method, RESOURCE_TEMPLATE);
-        for (MethodParameterInfo param : parameters) {
-            String paramName = param.name();
-            AnnotationInstance resourceTemplateArg = param.annotation(DotNames.RESOURCE_TEMPLATE_ARG);
-            if (resourceTemplateArg != null) {
-                AnnotationValue nameValue = resourceTemplateArg.value("name");
-                if (nameValue != null && !ResourceTemplateArg.ELEMENT_NAME.equals(nameValue.asString())) {
-                    paramName = nameValue.asString();
-                }
-            }
-            if (!param.type().name().equals(DotNames.STRING)) {
-                throw new IllegalStateException(
-                        "Resource template method must only consume String parameters: " + methodDesc(method));
-            }
-            if (!variableMatcher.variables().contains(paramName)) {
-                throw new IllegalStateException(
-                        "Parameter [" + paramName + "] does not match an URI template variable: "
-                                + methodDesc(method));
-            }
-        }
-    }
-
-    private void validateNotificationMethod(MethodInfo method) {
-        if (method.returnType().kind() != Kind.VOID
-                && (!method.returnType().name().equals(DotNames.UNI)
-                        || !method.returnType().asParameterizedType().arguments().get(0).name()
-                                .equals(DotName.createSimple(Void.class)))) {
-            throw new IllegalStateException("Notification method must return void or Uni<Void>");
-        }
-        List<MethodParameterInfo> params = parameters(method, NOTIFICATION);
-        if (!params.isEmpty()) {
-            throw new IllegalStateException(
-                    "Notification method %s may not consume the following parameter types: %s".formatted(
-                            methodDesc(method),
-                            params.stream().map(MethodParameterInfo::type).toList()));
-        }
-    }
-
-    private List<MethodParameterInfo> parameters(MethodInfo method, Feature feature) {
-        List<MethodParameterInfo> ret = new ArrayList<>();
-        for (MethodParameterInfo param : method.parameters()) {
-            Provider provider = providerFrom(param.type());
-            if (!provider.isValidFor(feature)) {
-                throw new IllegalStateException(
-                        "%s feature method %s may not accept parameter of type %s".formatted(feature,
-                                methodDesc(method), param.type()));
-            }
-            if (provider == Provider.PARAMS) {
-                ret.add(param);
-            }
-        }
-        return ret;
     }
 
     private MethodDesc processFeatureMethod(AtomicInteger counter, ClassCreator cc,
@@ -1659,16 +1324,16 @@ class McpServerProcessor {
                             if (requiredValue != null) {
                                 // Annotation value always takes precedence
                                 required = requiredValue.asBoolean();
-                            } else if (isOptionalType(param.type().name())
-                                    || hasDefaultValue(param)) {
+                            } else if (FeatureArguments.isOptionalType(param.type().name())
+                                    || FeatureArguments.hasDefaultValue(param)) {
                                 // No annotation value is defined and Optional type is used or default value set
                                 required = false;
                             }
-                            AnnotationValue defaultValueValue = argAnnotation.value(DEFAULT_VALUE);
+                            AnnotationValue defaultValueValue = argAnnotation.value(FeatureMethods.DEFAULT_VALUE);
                             if (defaultValueValue != null) {
                                 defaultValue = defaultValueValue.asString();
                             }
-                        } else if (isOptionalType(param.type().name())) {
+                        } else if (FeatureArguments.isOptionalType(param.type().name())) {
                             // No annotation is defined and Optional type is used
                             required = false;
                         }
@@ -1695,7 +1360,7 @@ class McpServerProcessor {
                             Const.of(required),
                             type,
                             defaultValue != null ? Const.of(defaultValue) : Const.ofNull(String.class),
-                            Const.of(providerFrom(param.type()))));
+                            Const.of(FeatureArguments.providerFrom(param.type()))));
                     bc.withList(args).add(arg);
                 }
 
@@ -1822,65 +1487,6 @@ class McpServerProcessor {
         });
     }
 
-    private boolean hasDefaultValue(MethodParameterInfo param) {
-        AnnotationInstance anno = param.annotation(DotNames.TOOL_ARG);
-        if (anno == null) {
-            anno = param.annotation(DotNames.PROMPT_ARG);
-        }
-        if (anno == null) {
-            anno = param.annotation(DotNames.MCPJAVA_TOOL_ARG);
-        }
-        if (anno == null) {
-            anno = param.annotation(DotNames.MCPJAVA_PROMPT_ARG);
-        }
-        return anno != null && anno.value(DEFAULT_VALUE) != null;
-    }
-
-    private FeatureArgument.Provider providerFrom(org.jboss.jandex.Type type) {
-        if (type.name().equals(DotNames.MCP_CONNECTION)) {
-            return FeatureArgument.Provider.MCP_CONNECTION;
-        } else if (type.name().equals(DotNames.REQUEST_ID)) {
-            return FeatureArgument.Provider.REQUEST_ID;
-        } else if (type.name().equals(DotNames.MCP_LOG)) {
-            return FeatureArgument.Provider.MCP_LOG;
-        } else if (type.name().equals(DotNames.REQUEST_URI)) {
-            return FeatureArgument.Provider.REQUEST_URI;
-        } else if (type.name().equals(DotNames.PROGRESS)) {
-            return FeatureArgument.Provider.PROGRESS;
-        } else if (type.name().equals(DotNames.ROOTS)) {
-            return FeatureArgument.Provider.ROOTS;
-        } else if (type.name().equals(DotNames.SAMPLING)) {
-            return FeatureArgument.Provider.SAMPLING;
-        } else if (type.name().equals(DotNames.CANCELLATION)) {
-            return FeatureArgument.Provider.CANCELLATION;
-        } else if (type.name().equals(DotNames.RAW_MESSAGE)) {
-            return FeatureArgument.Provider.RAW_MESSAGE;
-        } else if (type.name().equals(DotNames.COMPLETE_CONTEXT)) {
-            return FeatureArgument.Provider.COMPLETE_CONTEXT;
-        } else if (type.name().equals(DotNames.META)) {
-            return FeatureArgument.Provider.META;
-        } else if (type.name().equals(DotNames.ELICITATION)) {
-            return FeatureArgument.Provider.ELICITATION;
-        } else if (type.name().equals(DotNames.MCPJAVA_PROGRESS)) {
-            return FeatureArgument.Provider.MCPJAVA_PROGRESS;
-        } else if (type.name().equals(DotNames.MCPJAVA_CANCELLATION)) {
-            return FeatureArgument.Provider.MCPJAVA_CANCELLATION;
-        } else if (type.name().equals(DotNames.MCPJAVA_MCP_REQUEST)) {
-            return FeatureArgument.Provider.MCPJAVA_MCP_REQUEST;
-        } else if (type.name().equals(DotNames.MCPJAVA_COMPLETION_CONTEXT)) {
-            return FeatureArgument.Provider.MCPJAVA_COMPLETION_CONTEXT;
-        } else {
-            return FeatureArgument.Provider.PARAMS;
-        }
-    }
-
-    private static boolean isOptionalType(DotName name) {
-        return DotNames.OPTIONAL.equals(name)
-                || DotNames.OPTIONAL_INT.equals(name)
-                || DotNames.OPTIONAL_LONG.equals(name)
-                || DotNames.OPTIONAL_DOUBLE.equals(name);
-    }
-
     private Var getMapper(BlockCreator bc, FeatureMethodBuildItem featureMethod) {
         // Returns a function that converts the returned object to Uni<RESPONSE>
         // where the RESPONSE is one of ToolResponse, PromptResponse, ResourceResponse,
@@ -1907,7 +1513,8 @@ class McpServerProcessor {
                     DotNames.UNI.equals(returnType.name()) ? "CompleteUniOfMcpJavaResult" : "CompleteOfMcpJavaResult");
         }
         return readResultMapper(bc,
-                createMapperClassSimpleName(PROMPT_COMPLETE, returnType, DotNames.COMPLETE_RESPONSE, c -> "String"));
+                ResultMapperNames.createMapperClassSimpleName(PROMPT_COMPLETE, returnType, DotNames.COMPLETE_RESPONSE,
+                        c -> "String"));
     }
 
     Var resourceResultMapper(FeatureMethodBuildItem featureMethod, BlockCreator bc,
@@ -1920,11 +1527,12 @@ class McpServerProcessor {
                     DotNames.UNI.equals(returnType.name()) ? "ResourceUniOfMcpJavaResponse"
                             : "ResourceOfMcpJavaResponse");
         }
-        if (useEncoder(returnType, RESOURCE_TYPES)) {
+        if (useEncoder(returnType, FeatureMethods.RESOURCE_TYPES)) {
             return encoderResultMapper(featureMethod, bc, returnType, ResourceContentsEncoderResultMapper.class);
         } else {
             return readResultMapper(bc,
-                    createMapperClassSimpleName(RESOURCE, returnType, DotNames.RESOURCE_RESPONSE, c -> "Content"));
+                    ResultMapperNames.createMapperClassSimpleName(RESOURCE, returnType, DotNames.RESOURCE_RESPONSE,
+                            c -> "Content"));
         }
     }
 
@@ -1962,169 +1570,35 @@ class McpServerProcessor {
 
     Var toolResultMapper(FeatureMethodBuildItem featureMethod, BlockCreator bc,
             org.jboss.jandex.Type returnType) {
-        if (useEncoder(returnType, TOOL_TYPES)) {
+        if (useEncoder(returnType, FeatureMethods.TOOL_TYPES)) {
             if (featureMethod.isStructuredContent()) {
                 return encoderResultMapper(featureMethod, bc, returnType, ToolStructuredContentResultMapper.class);
             } else {
                 return encoderResultMapper(featureMethod, bc, returnType, ToolEncoderResultMapper.class);
             }
         } else {
-            return readResultMapper(bc, createMapperClassSimpleName(TOOL, returnType, DotNames.TOOL_RESPONSE, c -> {
-                return isContent(c) ? "Content" : "String";
-            }));
+            return readResultMapper(bc, ResultMapperNames.createMapperClassSimpleName(TOOL, returnType,
+                    DotNames.TOOL_RESPONSE, c -> {
+                        return ResultMapperNames.isContent(c) ? "Content" : "String";
+                    }));
         }
     }
 
     Var promptResultMapper(FeatureMethodBuildItem featureMethod, BlockCreator bc,
             org.jboss.jandex.Type returnType) {
-        if (useEncoder(returnType, PROMPT_TYPES)) {
+        if (useEncoder(returnType, FeatureMethods.PROMPT_TYPES)) {
             return encoderResultMapper(featureMethod, bc, returnType, PromptEncoderResultMapper.class);
         } else {
             return readResultMapper(bc,
-                    createMapperClassSimpleName(PROMPT, returnType, DotNames.PROMPT_RESPONSE, c -> {
+                    ResultMapperNames.createMapperClassSimpleName(PROMPT, returnType, DotNames.PROMPT_RESPONSE, c -> {
                         return DotNames.STRING.equals(c) ? "String" : "OfMessage";
                     }));
         }
     }
 
-    static String createMapperClassSimpleName(Feature feature, org.jboss.jandex.Type returnType,
-            DotName baseType, Function<DotName, String> componentMapper) {
-        if (returnType.name().equals(baseType)) {
-            return "ToUni";
-        }
-        org.jboss.jandex.Type type = returnType;
-        StringBuilder ret;
-        if (feature == PROMPT_COMPLETE || feature == RESOURCE_TEMPLATE_COMPLETE) {
-            ret = new StringBuilder("Complete");
-        } else {
-            String f = feature.toString();
-            // TOOL -> Tool
-            ret = new StringBuilder().append(f.charAt(0)).append(f.substring(1).toLowerCase());
-        }
-        if (DotNames.UNI.equals(type.name())) {
-            type = type.asParameterizedType().arguments().get(0);
-            if (type.name().equals(baseType)) {
-                return "Identity";
-            }
-            ret.append("Uni");
-        }
-        if (DotNames.LIST.equals(type.name())) {
-            type = type.asParameterizedType().arguments().get(0);
-            ret.append("List");
-        }
-        ret.append(componentMapper.apply(type.name()));
-
-        return ret.toString();
-    }
-
-    private boolean isContent(DotName typeName) {
-        return DotNames.CONTENT.equals(typeName)
-                || DotNames.TEXT_CONTENT.equals(typeName)
-                || DotNames.IMAGE_CONTENT.equals(typeName)
-                || DotNames.EMBEDDED_RESOURCE.equals(typeName)
-                || DotNames.RESOURCE_LINK.equals(typeName)
-                || DotNames.AUDIO_CONTENT.equals(typeName);
-    }
-
     private Var readResultMapper(BlockCreator bc, String mapperClassSimpleName) {
         ClassDesc mapperClassDesc = ClassDesc.of(ResultMappers.class.getName() + "$" + mapperClassSimpleName);
         return Expr.staticField(FieldDesc.of(mapperClassDesc, "INSTANCE", mapperClassDesc));
-    }
-
-    private static ExecutionModel executionModel(MethodInfo method,
-            TransformedAnnotationsBuildItem transformedAnnotations) {
-        if (KotlinUtils.isKotlinSuspendMethod(method)
-                && (transformedAnnotations.hasAnnotation(method, DotNames.RUN_ON_VIRTUAL_THREAD)
-                        || transformedAnnotations.hasAnnotation(method.declaringClass(),
-                                DotNames.RUN_ON_VIRTUAL_THREAD)
-                        || transformedAnnotations.hasAnnotation(method, DotNames.BLOCKING)
-                        || transformedAnnotations.hasAnnotation(method, DotNames.NON_BLOCKING))) {
-            throw new IllegalStateException("Kotlin `suspend` functions in MCP components may not be "
-                    + "annotated @Blocking, @NonBlocking or @RunOnVirtualThread: " + methodDesc(method));
-        }
-        if (transformedAnnotations.hasAnnotation(method, DotNames.RUN_ON_VIRTUAL_THREAD)
-                || transformedAnnotations.hasAnnotation(method.declaringClass(), DotNames.RUN_ON_VIRTUAL_THREAD)) {
-            return ExecutionModel.VIRTUAL_THREAD;
-        } else if (transformedAnnotations.hasAnnotation(method, DotNames.BLOCKING)) {
-            return ExecutionModel.WORKER_THREAD;
-        } else if (transformedAnnotations.hasAnnotation(method, DotNames.NON_BLOCKING)) {
-            return ExecutionModel.EVENT_LOOP;
-        } else if (transformedAnnotations.hasAnnotation(method, DotNames.TRANSACTIONAL)
-                || transformedAnnotations.hasAnnotation(method.declaringClass(), DotNames.TRANSACTIONAL)) {
-            // Method annotated with @Transactional or declared on a class annotated
-            // @Transactional is also treated as a blocking method
-            return ExecutionModel.WORKER_THREAD;
-        } else {
-            return hasBlockingSignature(method) ? ExecutionModel.WORKER_THREAD : ExecutionModel.EVENT_LOOP;
-        }
-    }
-
-    static boolean hasBlockingSignature(MethodInfo method) {
-        if (KotlinUtils.isKotlinSuspendMethod(method)) {
-            return false;
-        }
-        switch (method.returnType().kind()) {
-            case VOID:
-            case CLASS:
-            case ARRAY:
-                return true;
-            case PARAMETERIZED_TYPE:
-                // Uni, Multi -> non-blocking
-                DotName name = method.returnType().asParameterizedType().name();
-                return !name.equals(DotNames.UNI) && !name.equals(DotNames.MULTI);
-            default:
-                throw new IllegalStateException(
-                        "Unsupported return type:" + methodDesc(method));
-        }
-    }
-
-    private static String methodDesc(MethodInfo method) {
-        StringBuilder builder = new StringBuilder()
-                .append(method.declaringClass().name().withoutPackagePrefix())
-                .append("#")
-                .append(method.name())
-                .append('(');
-        for (Iterator<org.jboss.jandex.Type> it = method.parameterTypes().iterator(); it.hasNext();) {
-            builder.append(it.next());
-            if (it.hasNext()) {
-                builder.append(", ");
-            }
-        }
-        builder.append(')');
-        return builder.toString();
-    }
-
-    private static boolean isReturnTypeReflectionNeeded(DotName returnTypeName) {
-        return !isNameInTypes(returnTypeName, TOOL_TYPES)
-                && !isNameInTypes(returnTypeName, PROMPT_TYPES)
-                && !isNameInTypes(returnTypeName, RESOURCE_TYPES)
-                && !isNameInTypes(returnTypeName, COMPLETE_TYPES);
-    }
-
-    private static boolean isNameInTypes(DotName name, Set<org.jboss.jandex.Type> types) {
-        for (org.jboss.jandex.Type type : types) {
-            if (name.equals(type.name())) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    private static boolean isParamTypeReflectionNeeded(org.jboss.jandex.Type paramType) {
-        return paramType.kind() != Kind.PRIMITIVE
-                && !paramType.name().equals(DotNames.STRING)
-                && !paramType.name().equals(DotNames.MCP_CONNECTION)
-                && !paramType.name().equals(DotNames.MCP_LOG)
-                && !paramType.name().equals(DotNames.REQUEST_ID)
-                && !paramType.name().equals(DotNames.REQUEST_URI)
-                && !paramType.name().equals(DotNames.PROGRESS)
-                && !paramType.name().equals(DotNames.ROOTS)
-                && !paramType.name().equals(DotNames.SAMPLING)
-                && !paramType.name().equals(DotNames.CANCELLATION)
-                && !paramType.name().equals(DotNames.COMPLETE_CONTEXT)
-                && !paramType.name().equals(DotNames.RAW_MESSAGE)
-                && !paramType.name().equals(DotNames.META)
-                && !paramType.name().equals(DotNames.ELICITATION);
     }
 
 }
