+++ core/deployment/src/test/java/io/quarkiverse/mcp/server/deployment/McpServerProcessorTest.java
@@ -13,12 +13,12 @@ public class McpServerProcessorTest {
     @Test
     public void testCreateMapperField() {
         assertEquals("ResourceContent",
-                McpServerProcessor.createMapperClassSimpleName(Feature.RESOURCE,
+                ResultMapperNames.createMapperClassSimpleName(Feature.RESOURCE,
                         ClassType.create(DotNames.TEXT_RESOURCE_CONTENTS),
                         DotNames.RESOURCE_RESPONSE,
                         c -> "Content"));
         assertEquals("Identity",
-                McpServerProcessor.createMapperClassSimpleName(Feature.RESOURCE,
+                ResultMapperNames.createMapperClassSimpleName(Feature.RESOURCE,
                         ParameterizedType.create(DotNames.UNI, ClassType.create(DotNames.RESOURCE_RESPONSE)),
                         DotNames.RESOURCE_RESPONSE,
                         c -> "Content"));
